move types to a separate file that can be included from lantern.
